modified audioengine presetnum max, modified random trigger (old was phase unprocessed, now is phase for each oscillator)

+++ src/parametersControl.cpp
@@ -847,7 +847,7 @@ void parametersControl::loadPreset(string presetName, string bank){
         }
         
         //if we have to destroy the module we do it
-        if(hasToBeDestroyed && moduleName != "senderManager" && moduleName != "waveScope" && moduleName != "typeConverter" && moduleName != "audioControls"){
+        if(hasToBeDestroyed && moduleName != "senderManager" && moduleName != "waveScope" && moduleName != "typeConverter" && moduleName != "audioControls" && moduleName != "colorApplier"){
             destroyModuleAndConnections(i);
         }
         else{
